requirement.txt
# Data models
pydantic>=2.0

# Web framework + server
fastapi>=0.110
uvicorn>=0.30

# Kafka streaming (Phase 2)
kafka-python>=2.0

# Exchange API clients & Real-time streaming
requests>=2.31
websocket-client>=1.8.0

# Testing & linting
pytest>=8.0
ruff>=0.5

# Interactive terminal UI
textual>=0.61
